[MID-6] changed files structure
