+++ src/app/api/admin/[...slug]/consts.js
@@ -1,0 +1,4 @@
+export const MODEL_SEARCH_FIELDS = {
+  Specialist: ['firstName', 'lastName', 'surname'],
+  Organization: ['name'],
+};
