+++ src/app/api/admin/[...slug]/inputFilters.js
@@ -1,0 +1,7 @@
+import { MODEL_SEARCH_FIELDS } from './consts';
+
+export function searchInputFilters(modelName, filter) {
+  if (!filter) return {};
+  const filters = MODEL_SEARCH_FIELDS[modelName].map(field => ({ [field]: { contains: filter, mode: 'insensitive' } }));
+  return { OR: filters };
+}
