+++ src/lib/consts.js
@@ -10,10 +10,7 @@ export const BASE_ERROR_MESSAGES = {
   429: 'Too Many Requests',
   500: 'Internal Server Error',
 };
-export const MODEL_SEARCH_FIELDS = {
-  Specialist: ['firstName', 'lastName', 'surname'],
-  Organization: ['name'],
-};
+
 export const MODEL_INCLUDES = {
   Specialist: {
     therapies: { select: { name: true } },
